Improved callback on successful or failed transaction that returns result status and BoC and more
Added additional check for matching sender and receiver addresses to cancel the transaction in case of a match.
Improved callback on successful or failed transaction that returns result status and BoC.
Improved implementation of SDK initialization and transaction status handling in the demo for reference.

Removed the ZXing library for generating QR codes on .NET as it is not used in the SDK.